refactor: clean up unnecessary in the mashup example
