--- mashup-example/Qlik-Sense-Desktop/Using-Build-Tool/.gitignore
@@ -1,24 +1,0 @@
-*.log
-*.rej
-*.tmp
-*.log
-.cache/
-.DS_Store
-.idea/
-.vscode/
-.npmrc
-.parcel-cache
-
-node_modules/
-coverage/
-dist/
-temp/
-test/**/__artifacts__/regression
-test/**/__artifacts__/diff
-apis/*/core/**/*.js
-apis/*/core/**/*.js.map
-apis/locale/all.json
-
-Search/
-
-local-dev/
