--- mashup-example/Qlik-Sense-SaaS/.gitignore
@@ -1,11 +1,0 @@
-*.log
-*.log
-.cache
-.DS_Store
-.idea
-.vscode
-.npmrc
-
-node_modules
-coverage
-dist/
