+++ mashup-example/qlik-sense-desktop/using-build-tool/.gitignore
@@ -1,0 +1,1 @@
+.parcel-cache
