+++ mashup-example/qlik-sense-desktop/using-build-tool/connect.js
