+++ mashup-example/qlik-sense-saas/.gitignore
@@ -1,0 +1,1 @@
+.parcel-cache
