+++ mashup-example/qlik-sense-saas/package.json
