+++ mashup-example/qlik-sense-saas/src/configure.js
