+++ mashup-example/qlik-sense-saas/src/connect.js
