+++ mashup-example/qlik-sense-saas/src/index.js
